Аудит Stonks: целостность сделок, точные деньги, лимиты API (#118)
* perf: загружать котировки списком и кэшировать на 5 секунд

Одна команда пользователя пересчитывала портфель по нескольку раз
(recordBaseline → equity, затем grossPositionValue отдельно, затем
проверка риска), и каждый шаг делал отдельный сетевой вызов к API
T-Банка на каждую позицию. Планировщик заявок раз в минуту делал
по запросу на каждую заявку, хотя API принимает список инструментов.

SandboxPriceService получил loadPrices(Collection) и кэш на 5 секунд
(меньше периода планировщика). На батч переведены grossPositionValue,
portfolio, liquidate и все три проверки в SandboxOrderProcessor.

Попутно:
- комиссия считается от точного turnover, а не через
  BigDecimal.valueOf(price.doubleValue()) — прогон через double терял
  точность ровно там, где комментарий обещал её сохранить;
- в liquidate комиссия округляется до копеек, как в обычной сделке
  (было scale=8), а стоп-заявки удаляются точечным
  deleteByUserAndTicker вместо перебора всех заявок всех пользователей.


* fix: сделка выполняется одной транзакцией, ошибки БД больше не глотаются

1. Все семь репозиториев песочницы оборачивали каждый запрос в
   try/catch (Exception) с логом и возвращали void/null/пустой список.
   Последствия: при сбое БД positions.save() «проходил» молча, а
   users.save() ниже отрабатывал — бумаги без списания денег; findById()
   возвращал null, и бот отвечал «Сначала выполните +регистрация»;
   findAll() отдавал пустоту, из-за чего процессор решал, что стоп-лоссов
   нет, и не исполнял их. Перехват убран — DataAccessException доходит
   до вызывающего.

2. Сделка (позиция + счёт + запись в историю + возможная ликвидация)
   выполняется одной транзакцией через TransactionTemplate. Откат по
   риску теперь делает сама транзакция, ручной rollbackTrade удалён.
   DM о margin call отправляется после коммита, а не внутри транзакции.

3. serializeHoldings при ошибке подставлял "{}" — валютные остатки
   молча обнулялись в БД. Теперь бросает, транзакция откатывается.
   Чтение повреждённого JSON по-прежнему не роняет пользователя целиком.

4. MessageHandler при ошибке команды только писал в лог — пользователь
   не понимал, прошла операция или нет. Теперь отвечает в канал.


* fix: порог margin call берётся из конфигурации

sandbox.maintenance-margin (0.25) нигде в логике не использовался —
реальный порог был зашит константой 0.2 в SandboxRiskManager. При этом
+маржа печатала настройку как «Порог margin call: 0.25», а строкой ниже
«Порог ликвидации: 0.20» — то есть показывала настоящий порог под чужой
подписью, а настраиваемое значение не влияло ни на что.

Теперь RiskManager сравнивает margin level с maintenanceMargin, лишняя
строка из +маржа убрана. Порог стал строже (0.25 вместо 0.2) — так было
описано в конфиге и в ответе бота.


* refactor: деньги и даты хранятся точными типами

Балансы, средние цены и комиссии были double в моделях и
DOUBLE PRECISION в схеме — копейки дрейфовали при каждом накоплении,
а сравнения сумм зависели от порядка операций. Даты (baseline-даты,
last_replenish_date) хранились строками.

- модели: cash, borrowed, total_fees, avg_price, price, fee,
  limit_price, trigger_price, target_price, amount_rub и валютные
  остатки — BigDecimal;
- схема: NUMERIC(19,8) для денег, DATE для дат, с идемпотентной
  миграцией ALTER ... USING в schema.sql;
- репозитории читают getBigDecimal/getObject(LocalDate) вместо
  getDouble и ручного LocalDate.parse;
- BaseRepository.nz() приводит NULL к нулю: getDouble() возвращал 0.0
  сам, getBigDecimal() отдаёт null, и старая запись с пустым полем
  иначе роняла бы расчёты.

Проверено на копии боевой БД, поднятой из свежего pg_dump: типы
сменились, балансы, позиции и сделки на месте, повторный прогон
schema.sql проходит без ошибок.

ВАЖНО: после миграции старый jar с БД не работает — откат деплоя
потребует восстановления дампа
(/mnt/storage/backups/postgres/stonks_pre_migration_20260902_1151.sql).


* fix: текст команды +дайджест соответствует реальному расписанию

После перевода дайджеста на еженедельный режим (#117) команда
+дайджест вкл продолжала обещать «Каждый день в 9:00».


---------

+++ src/main/java/services/sandbox/DcaScheduler.java
@@ -69,7 +69,7 @@ public class DcaScheduler {
                 return;
             }
 
-            BigDecimal amount = BigDecimal.valueOf(order.getAmountRub());
+            BigDecimal amount = order.getAmountRub();
             int qty = amount.divide(price, 0, RoundingMode.DOWN).intValue();
             if (qty <= 0) {
                 log.warn("DCA: недостаточно суммы {} для покупки {} (цена {}), пропускаем id={}",
